+++ packages/widgetbook/lib/src/addons/device_frame_addon/addon.dart
@@ -2,3 +2,4 @@ export 'package:device_frame/device_frame.dart';
 
 export 'device_frame_addon.dart';
 export 'device_frame_setting.dart';
+export 'none_device.dart';