+++ packages/widgetbook/lib/src/addons/device_frame_addon/device_frame_setting.dart
@@ -8,7 +8,7 @@ class DeviceFrameSetting {
     this.hasFrame = true,
   });
 
-  final DeviceInfo? device;
+  final DeviceInfo device;
   final Orientation orientation;
   final bool hasFrame;
 }